Moves and updates accessibility statement
Moves the accessibility statement to a more appropriate directory and updates its content to provide more comprehensive information.

The statement now includes details on Coello One's commitment to accessibility, contact information for reporting issues, and a disclaimer regarding legal rights.

Additionally, updates layout of other pages and adds a back button to the home page. Updates the accessibility statement link in the sider footer
