+++ coello-one/app/[locale]/(other-pages)/blog/page.tsx
